requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.7.0
python-dotenv==1.1.0
psycopg2==2.9.10
python-socketio==5.13.0
uvicorn==0.34.2
requests==2.32.3
django-environ==0.12.0
django-guest-user==0.5.5
